+++ lib/util/text_damage_component.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:bonfire/base/rpg_game.dart';
+import 'package:bonfire/base/bonfire_game.dart';
 import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:flame/components/mixins/has_game_ref.dart';
@@ -8,7 +8,7 @@ import 'package:flame/text_config.dart';
 
 enum DirectionTextDamage { LEFT, RIGHT, RANDOM, NONE }
 
-class TextDamageComponent extends TextComponent with HasGameRef<RPGGame> {
+class TextDamageComponent extends TextComponent with HasGameRef<BonfireGame> {
   final String text;
   final TextConfig config;
   final Position initPosition;