+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # NEXT
 
 - BREAKING CHANGE: Collision system. Remove param `collision` from Enemy, Player and GameDecoration. If you need add collision in your component
-use the mixin 'Collision' and settings your properties using 'setupCollision()' method.
+use the mixin 'Collision' and settings your properties using 'setupCollision()' method. See [example](https://github.com/RafaelBarbosatec/bonfire/blob/master/example/lib/enemy/goblin.dart).
+-  Improvements in `DragGesture` and `TapGesture`
 
 # 0.8.6